Added validation for bools. Refined the test structure: moved the validation code into the type to save typing in the actual validation methods that users would have to write (or would be covered by the reflection).
Created a range object for int's. This enforced the ordering of values. Created an exclusive one which warns developer if the start and end values are equal.

Created the slice object to test slice properties. still working on how to test the contents of the slice and how to make slice packaging easier. Worked around the inability to cast slice of type by wrapping the object in a struct and providing a Len()int method.